LU-15458 tests: double delay for sync/async test
This test verifies that an async request takes less than
half of the 'delay' time, but occasionally the async
request hits a delay on the file open due to some VM issue
on the MDT.  This causes a spurious failure.

This is already pretty rare, so the simplest thing is to
double the delay time, which should make it extremely rare.

Setting trivial so it just runs sanity.

test-parameters: trivial

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: I3046303fcc4e10364de9f673fab2142c8cecff64
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/51183
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>