LU-9859 lnet: simplify cfs_parse_nidlist()
By duplicating the string being parsed, we can use mutating parsign
primitives and simplify the code.

Note that the kernel-space cfs_parse_nidlist() is now different from
the user-space version.  As they are both declared in the same header
file, this needs an #ifdef until the headers can be separated.

Change some function that return 0 on error to match the kernel
convention of 0 on success and -ve on error.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I1d5155a1cee82f798bec0863d80d75af92399cf1
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/50841
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>