LU-15367 tests: Multiop command from file
Multiop has the potential to be really useful as an I/O
pattern simulator, in that it can perform almost any
operation.  In theory, we could process iotrace logs
from a user application, and then generate command
strings for multiop.  This works great, but we very
quickly run in to the limit of the maximum command input
size (the shell limit).

So, add the ability for multiop to take command input from
a file.  This is implemented a little awkwardly in order to
make it entirely optional and not change the many existing
tests using multiop.

Test-Parameters: trivial
Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: If3da66a4d1c4ab321f8381163c6ac4aecbe61923
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/45840
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Tested-by: Maloo <maloo@whamcloud.com>