LU-18808 lnet: allow LNDs to discover metadata via pings
The EFA LND needs to be able to automatically discover remote
EFA NIs GID and manager QP data by first performing a handshake
using LNet pings over TCP. This allows communication over EFA
without relying on large NID support and without needing the GID to
be provided externally.

The metadata for all remote NIs for a particular node are passed
via the LNet ping REPLY packet. Since a node will only send its
own NIs metadata, a handshake must be performed with each node
in a cluster.

We define a new LNet callback lnd_get_nid_metadata to enable
LNet to query the LND for NID related metadata to send
alongside the ping REPLY. Of course, this is implemented by the
EFA LND to transmit the GIDs and manager QP data for local NIs.

Currently, this functionality is not used. The first consumer
will be the EFA LND after it is upstreamed.

Test-Parameters: trivial testlist=sanity-lnet serverversion=2.16
Signed-off-by: Timothy Day <timday@amazon.com>
Change-Id: I8431da6296705ee4fd64cebf2a9f218cafbfa1a8
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58407
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Yonatan Nachum <ynachum@amazon.com>
Reviewed-by: Frank Sehr <fsehr@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>